+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Version 1.2
 ** Other Major changes **
 
 * Refactor of the action provider, so that list of available actions only contains the title, name and not the instaciated class (with all the loaded configuration, such as groups, option values etc..)
+* Create Address actions now also contains parameters for street name, housenumber and housenumber suffix
 
 ** New actions**